refactor: extract ScreenshotCaptureSupport.swift (4 helper types)
Extracts 4 supporting types from ScreenshotCaptureService.swift into a new
ScreenshotCaptureSupport.swift:
- struct ScreenCaptureDisplaySnapshot (7-13)
- struct CapturedDisplayImage (14-28)
- struct ScreenCaptureKitImageProvider (30-143) — implements ScreenCaptureImageProviding
- struct PNGScreenshotImageWriter (145-186) — implements ScreenshotImageWriting

Byte-identical move; SHA-256 851342564a2758b93df66ce7a92f8013a5bc0a253637869675f27709db7561c3
verified against origin/main:ScreenshotCaptureService.swift lines 7-186.

Imports: CoreGraphics + Foundation + @preconcurrency ScreenCaptureKit.

Reduces ScreenshotCaptureService.swift from 385 to 200 lines. The primary
ScreenshotCaptureService struct + trailing helpers remain.

Tests: 667 (unchanged).
